requirements.txt
ftfy==6.1.1
huggingface-hub==0.31.2
numpy==1.24.4
opencv-python==4.6.0.66
pandas==2.2.3
Pillow==9.2.0
regex==2022.10.31
scikit-image==0.19.3
scikit-learn==1.6.1
setuptools==63.4.1
tabulate==0.9.0
timm==0.6.12
torch==2.7.1
torchaudio==2.7.1
torchdata==0.11.0
torchmetrics==0.10.3
torchsummary==1.5.1
torchvision==0.22.1
tqdm==4.65.0
matplotlib==3.9.4